3__crud_rest_validation: handling exceptions and validation

+++ 3__rest_crud_validation/src/main/java/com/devsuperior/rest_crud/services/ClientService.java
@@ -2,7 +2,9 @@ package com.devsuperior.rest_crud.services;
 
 import com.devsuperior.rest_crud.dto.ClientDTO;
 import com.devsuperior.rest_crud.entities.Client;
+import com.devsuperior.rest_crud.exceptions.ResourceNotFoundException;
 import com.devsuperior.rest_crud.repositories.ClientRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -21,7 +23,7 @@ public class ClientService {
 
     @Transactional(readOnly = true)
     public ClientDTO findById(Long id){
-        return new ClientDTO(repository.findById(id).get());
+        return new ClientDTO(repository.findById(id).orElseThrow(()->new ResourceNotFoundException("Client not found")));
     }
 
     @Transactional
@@ -33,14 +35,20 @@ public class ClientService {
 
     @Transactional
     public ClientDTO update(Long id, ClientDTO dto){
-        Client client = repository.getReferenceById(id);
-        copyDTOtoEntity(client, dto);
-        return new ClientDTO(repository.save(client));
+        try{
+            Client client = repository.getReferenceById(id);
+            copyDTOtoEntity(client, dto);
+            return new ClientDTO(repository.save(client));
+        }catch(EntityNotFoundException ex){throw new ResourceNotFoundException("Client not found for update");}
     }
 
     @Transactional
     public void delete(Long id){
-        repository.deleteById(id);
+        if(!repository.existsById(id)){
+            throw new ResourceNotFoundException("Client not found for deletion");
+        }else{
+            repository.deleteById(id);
+        }
     }
 
     private void copyDTOtoEntity(Client entity, ClientDTO dto){
